// frontend/src/generated/client/types.gen.ts
export type ClientResponse = {
    /**
     * Id
     */
    id: string;
    /**
     * Organization Id
     */
    organization_id: string;
    /**
     * Name
     */
    name: string;
    /**
     * Code
     */
    code: string;
    /**
     * Description
     */
    description: string | null;
    /**
     * Status
     */
    status: 'active' | 'inactive';
    /**
     * Created At
     */
    created_at: string;
    /**
     * Updated At
     */
    updated_at: string;
};
// ...
export type ProjectResponse = {
    /**
     * Id
     */
    id: string;
    /**
     * Organization Id
     */
    organization_id: string;
    /**
     * Client Id
     */
    client_id: string;
    /**
     * Name
     */
    name: string;
    /**
     * Code
     */
    code: string;
    /**
     * Description
     */
    description: string | null;
    /**
     * Status
     */
    status: 'active' | 'inactive';
    /**
     * Created At
     */
    created_at: string;
    /**
     * Updated At
     */
    updated_at: string;
};